feat(sse)!: refactor cmd hierarchy to group transports under serve
