+++ cmd/root.go
@@ -22,3 +22,7 @@ func Execute() {
 		os.Exit(1)
 	}
 }
+
+func init() {
+	cobra.EnableTraverseRunHooks = true
+}
